Reduce duplication of credential manager code to make Sonar happy
