+++ packages/imperative/src/OverridesLoader.ts
@@ -15,6 +15,8 @@ import { IImperativeConfig } from "./doc/IImperativeConfig";
 import { isAbsolute, resolve } from "path";
 import { AppSettings } from "../../settings";
 import { ImperativeConfig } from "../../utilities";
+import { IConfigVault } from "../../config";
+import { Logger } from "../../logger";
 
 /**
  * Imperative-internal class to load overrides
@@ -64,6 +66,28 @@ export class OverridesLoader {
       // We do not use the invalid credential manager, since we no longer allow overrides.
       invalidOnFailure: false
     });
+
+    /**
+     * After the plugins and secure credentials are loaded, rebuild the configuration with the
+     * secure values
+     */
+    if (CredentialManagerFactory.initialized) {
+      const vault: IConfigVault = {
+        load: ((key: string): Promise<string> => {
+          return CredentialManagerFactory.manager.load(key, true);
+        }),
+        save: ((key: string, value: any): Promise<void> => {
+          return CredentialManagerFactory.manager.save(key, value);
+        })
+      };
+
+      try {
+        await ImperativeConfig.instance.config.secureLoad(vault);
+      } catch (err) {
+        // Secure vault is optional since we can prompt for values instead
+        Logger.getImperativeLogger().warn(`Secure vault not enabled. Reason: ${err.message}`);
+      }
+    }
   }
 
   /**
